+++ jest.config.js
@@ -3,6 +3,6 @@ const defaultConfig = require("./jest.config.standalone");
 // This config starts up nailgun by default
 module.exports = {
   ...defaultConfig,
-  globalSetup: "<rootDir>/tests_config/set_up.js",
-  globalTeardown: "<rootDir>/tests_config/tear_down.js",
+  globalSetup: "<rootDir>/bin/start-apex-server.js",
+  globalTeardown: "<rootDir>/bin/stop-apex-server.js",
 };